+++ HISTORY.md
@@ -1,3 +1,9 @@
+2.1.0 / 2020-11-27
+===================
+
+* Disable no-undef rule for TypeScript
+* Use TypeScript rules only for `**/*.ts` and `**/*.tsx`
+
 2.0.0 - 2.0.5 / 2020-10-02
 ===================